consumer/requirements.txt
confluent-kafka
telebot
kafka-python
pyTelegramBotAPI